requirements.txt
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
validate-docbr==1.10.0
